src/crisissum/backend/requirements.txt
faiss-gpu
sentence-transformers
fastapi
uvicorn
pandas
python-multipart
numpy==1.9